pyproject.toml
[dependency-groups]
dev = [
  "pytest",
  "ruff",
  "ty",
  "pre-commit",
  "python-dotenv"
]

[project]
dependencies = [
  "fastapi[all]>=0.124.4",
  "openai>=2.13.0",
  "pydantic>=2.12.5",
  "streamlit>=1.52.1",
  "authlib>=1.3.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0"
]
description = "Goatbot project"
name = "goatbot"
requires-python = ">=3.12"
version = "0.1.0"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
ignore = [
  "E501",  # line too long, handled by ruff format
  "B008",  # do not perform function calls in argument defaults
  "C901"  # too complex
]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP"  # pyupgrade
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"streamlit-chat/config.py" = ["E402"]  # import after sys.path modification is intentional
